Moved functions out of tweet_turing to twittertext
Moved four functions from tweet_turing.py, along with pre-req import statements for those functions.

These functions were used in 03_eda only, so this prevents subsequent notebooks from extra bloat when importing tweet_turing.py

Also modified 03_eda to reference the new source definition of these functions.
