+++ src/twittertext.py
@@ -4,14 +4,22 @@
 #   
 #   Original library: https://github.com/twitter/twitter-text/tree/master/java/src/main/java/com/twitter/twittertext
 #
-#   Also includes a function from external package `tweet_counter` to count the characters of a tweet.
+#   Also includes: 
+#    - a function from external package `tweet_counter` to count the characters of a tweet
+#    - a function from external package `tldextract` to extract top-level domain from a URL
 
 
 import re
+import pandas as pd
+import tldextract       # Source: https://github.com/john-kurkowski/tldextract
+
 from tweet_counter import count_tweet   
     # Source: https://github.com/nottrobin/tweet-counter
     #   Version: 0.1.0
     #   Note: on Windows, above package requires manual installation (`pip install` generates error)
+
+
+__all__ = ['char_count', 'retweet_handle', 'reply_handle', 'explode_url']
 
 
 # constants
@@ -60,14 +68,61 @@ VALID_REPLY_PATTERN = re.compile(VALID_REPLY_PATTERN_STR, flags=re.IGNORECASE)
 
 
 # functions
-def extract_reply_screenname(tweet_text: str):
+##########################
+#### PUBLIC INTERFACE ####
+##########################
+
+def char_count(tweet_series: pd.Series) -> int:
+    """Counts the number of characters in a tweet using an approximation of Twitter's specific counting method."""
+    return _get_tweet_char_count(tweet_series['content'])
+
+
+def retweet_handle(tweet_series: pd.Series) -> str:
+    """Returns the Twitter handle of the parent author of a retweeted tweet.
+        E.g. if @foo is retweeting a tweet by @bar with the text `RT @bar What is your name?`, returns `bar`."""
+    handle: str = _extract_first_handle_after_RT(tweet_series['content'])
+    if (handle is None):
+        return None
+    else:
+        return handle.lstrip('@')
+
+
+def reply_handle(tweet_series: pd.Series) -> str:
+    """Returns the Twitter handle of the parent author of a reply tweet.
+        E.g. if @foo is replying to a tweet by @bar with the text `@bar What is your quest?`, returns `bar`."""
+    handle: str = _extract_reply_screenname(tweet_series['content'])
+    if (handle is None):
+        return None
+    else:
+        return handle.lstrip('@')
+
+
+def explode_url(url_text: str) -> dict:
+    """Breaks apart a string containing only a fully-qualified URL.
+        Returns the pieces of the URL as a dict. Refer to tldextract docs
+        for more information."""
+    url_named_tuple: tldextract.ExtractResult = tldextract.extract(url_text)
+
+    return {
+            'subdomain': url_named_tuple.subdomain,
+            'domain': url_named_tuple.domain,
+            'tld': url_named_tuple.suffix,
+            'registered_domain': url_named_tuple.registered_domain
+        }
+
+
+###########################
+#### PRIVATE FUNCTIONS ####
+###########################
+
+def _extract_reply_screenname(tweet_text: str):
     """Returns the handle, including `@`, occuring at start of a tweet (i.e. when a tweet is replying to another).
         Returns None if anything but a handle is at the very start of a tweet.
         Regex based on Twitter's `twittertext` code."""
     return VALID_REPLY_PATTERN.match(tweet_text)    # seems to work, may consider "search" but "match" looks only at start of string
 
 
-def extract_first_handle_after_RT(tweet_text: str):
+def _extract_first_handle_after_RT(tweet_text: str):
     """Returns the handle, including `@`, occuring at start of a tweet but after chars 'RT ' (note the trailing space).
         Returns None if anything but 'RT ' followed by a handle is at the very start of a tweet.
         Note that check for 'RT ' is expected to be performed prior to calling this function.
@@ -75,7 +130,7 @@ def extract_first_handle_after_RT(tweet_text: str):
     return VALID_REPLY_PATTERN.search(tweet_text[3:])   # seems to work
 
 
-def get_tweet_char_count(tweet_text: str) -> int:
+def _get_tweet_char_count(tweet_text: str) -> int:
     """Counts the number of characters in a Tweet using an approximation of the method used by Twitter.
         Twitter determines character count using a very specific approach: https://developer.twitter.com/en/docs/counting-characters"""
     return count_tweet(tweet_text)
